moved community page to dashboard
